# Conflicts:
#	cloud_gateway/src/main/resources/application.yml
#	cloud_oauth2/src/main/resources/application.yml
#	cloud_provider/src/main/resources/application.yml
#	eureka-server/src/main/resources/application.yml

+++ cloud_provider/src/main/java/com/wangwei/cloud/sys/service/IUserRoleService.java
@@ -15,5 +15,5 @@ import java.util.Map;
  * @since 2020-03-23
  */
 public interface IUserRoleService extends IService<UserRole> {
-    public List<Map<String,Object>> getUserRoles(int userId );
+    public List<Map<String,Object>> getUserRoles(Long userId );
 }
